+++ phpunit/bootstrap.php
@@ -72,6 +72,7 @@ if (file_exists(GLPI_CONFIG_DIR . DIRECTORY_SEPARATOR . CacheManager::CONFIG_FIL
     $GLPI_CACHE = new SimpleCache(new ArrayAdapter());
 }
 
+# TODO: register a proper PSR4 autoloader for these files.
 include_once __DIR__ . '/GLPITestCase.php';
 include_once __DIR__ . '/DbTestCase.php';
 include_once __DIR__ . '/CsvTestCase.php';
@@ -84,6 +85,7 @@ include_once __DIR__ . '/abstracts/CommonITILRecurrentTest.php';
 include_once __DIR__ . '/AbstractRightsDropdown.php';
 include_once __DIR__ . '/CommonDropdown.php';
 include_once __DIR__ . '/HLAPITestCase.php';
+require_once __DIR__ . '/functional/Glpi/Form/Condition/ConditionHandler/AbstractConditionHandler.php';
 
 loadDataset();
