+++ phpunit/functional/Glpi/Form/Condition/EditorManagerTest.php
